Add unless clause

+++ modules/php5/manifests/extension/opcache.pp
@@ -28,6 +28,7 @@ class php5::extension::opcache (
       command     => 'find /etc/php5 -name 05-opcache.ini -delete',
       path        => ['/usr/local/sbin', '/usr/local/bin', '/usr/sbin', '/usr/bin', '/sbin', '/bin'],
       provider    => shell,
+      unless      => 'find /etc/php5/ -name 05-opcache.ini | grep -q 05-opcache.ini',
       refreshonly => true,
       subscribe   => Php5::Config_extension['opcache'],
     }
